refactor verification module to make pipeline clearer
