--- rapx/src/analysis/senryx/contracts/mod.rs
@@ -1,3 +1,0 @@
-pub mod abstract_state;
-pub mod property;
-// pub mod state_lattice;
