+++ rapx/src/analysis/verify/scan.rs
@@ -4,10 +4,7 @@ use rustc_hir::{
     def_id::{DefId, LocalDefId},
     intravisit::{FnKind, Visitor, walk_fn},
 };
-use rustc_middle::{
-    hir::nested_filter,
-    ty::TyCtxt,
-};
+use rustc_middle::{hir::nested_filter, ty::TyCtxt};
 use rustc_span::{Span, Symbol};
 
 /// Visitor that collects all functions annotated with `#[rapx::verify]`.
@@ -39,9 +36,7 @@ impl<'tcx> VerifyAttrVisitor<'tcx> {
 
             let path = attr.path(); // SmallVec<Symbol>
 
-            path.len() == 2
-                && path[0] == rapx
-                && path[1] == verify
+            path.len() == 2 && path[0] == rapx && path[1] == verify
         })
     }
 }
@@ -85,7 +80,10 @@ impl<'tcx> Analysis for VerifyScanAnalysis<'tcx> {
         rap_info!("======== #[rapx::verify] scan ========");
         let mut visitor = VerifyAttrVisitor::new(self.tcx);
         self.tcx.hir_visit_all_item_likes_in_crate(&mut visitor);
-        rap_info!("total: {} function(s) annotated with #[rapx::verify]", visitor.verify_fns.len());
+        rap_info!(
+            "total: {} function(s) annotated with #[rapx::verify]",
+            visitor.verify_fns.len()
+        );
         rap_info!("=====================================");
     }
