+++ rapx/tests/safety_check/align_cases/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "align_cases"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
